+++ mangadex_downloader/cli/args_parser.py
@@ -559,6 +559,10 @@ def get_args(argv):
 
     # Miscellaneous
     misc_group = parser.add_argument_group('Miscellaneous')
+    misc_group.add_argument(
+        '--input-pos',
+        help='Automatically select choices in selectable prompt (list, library, followed-list command)'
+    )
     misc_group.add_argument('-pipe', action='store_true', help="Download from pipe input")
     misc_group.add_argument(
         '--no-verify',