+++ mangadex_downloader/cli/download.py
@@ -28,15 +28,6 @@ from ..errors import MangaDexException
 log = logging.getLogger(__name__)
 
 def download(args):
-    # Single download
-    if len(args.URL) == 1:
-        url = args.URL[0]
-        
-        # Start downloading
-        url(args, args.type)
-        return
-    
-    # Multiple downloads
     for url in args.URL:
         try:
             url(args, args.type)